Dispose the proper orphan modules
When disposing orphans, make sure we are checking against the proper
moduleName. We were re-using the variable `moduleName` in a for loop
which caused the other moduleName to be changed, and as a result we were
disposing of some modules that we should not have been. This adds a test
and a fix.

Fixes #21