Adds a disposeModule method to the live-reload module
This adds a new function to the live-reload module that other modules
import. What this function does is perform the normal dispose behavior
of that module. This is useful in scenarios where a module is creating
virtual modules that will never be reloaded themselves, but still to
have dispose actions occur.

Fixes #22